Catalog template replacement - D3D9Catalog is now a template class   * no need for name anymore   * reduced memory footprint   * faster access @ Remove   * iterators (more C++ like ;) )   * takes advantage of better compiler optimizations   * added type-safety   * STL containers should be fairly thread-save - bumped version to Beta 23.3 - documentation updates
git-svn-id: svn://mirror.orbiter-radio.co.uk/D3D9client/trunk@625 c0f5cd98-1510-604a-8459-bf7538b8c9c2

+++ Orbitersdk/D3D9Client/Cloudmgr2.cpp
@@ -56,7 +56,7 @@ void CloudTile::Load ()
 			tex = getParent()->Tex();
 			owntex = false;
 		} else tex = 0;
-	} else TileCatalog->Add(DWORD(tex));
+	} else TileCatalog->Add(tex);
 
 	bool shift_origin = (lvl >= 4);
 	int res = mgr->Cprm().gridRes;
