+++ Orbitersdk/D3D9Client/Texture.cpp
@@ -109,7 +109,7 @@ int TextureManager::LoadTextures(const char *fname, LPDIRECT3DTEXTURE9 *ppdds, D
 			if (D3DXCreateTextureFromFileInMemoryEx(pDev, location, bytes, 0, 0, 1, 0, D3DFMT_FROM_FILE,
 				D3DPOOL_DEFAULT, D3DX_DEFAULT, D3DX_DEFAULT, 0, &Info, NULL, &pTex)==S_OK) {
 				ppdds[ntex] = pTex;
-				TileCatalog->Add(DWORD(pTex));
+				TileCatalog->Add(pTex);
 				//LogAlw("Loaded a texture from %s, 0x%X (%u x %u)", fname, pTex, Info.Width, Info.Height);
 			}
 			else {
